refactor for use on lambda (WIP)
